refactor crawler forms
